Changed the board to use 3.3v rather than 5v. Total new layout.
Since the FTDI chip cannot communicate with the 5v PIC. So run
the PIC at 3.3v This limits the maximum crystal frequency of
the pic, but 4MHz is fine. the internal usb regulator is no
longer used and is replaced with the external one.

The 3.3v regulator get it's 5v from the selected 5v source.
5v is still used by the FTDI chip and the Motor driver.

Signed-off-by: Kolijn Wolfaardt <kolijn.wolfaardt@gmail.com>